requirements.txt
selenium
requests
pytest
pytest-rerunfailures
pytest-xdist
allure-pytest
pyyaml
